+++ examples/complete/data.tf
@@ -19,6 +19,18 @@ data "aws_subnet" "public" {
   id       = each.value
 }
 
+data "aws_subnets" "private" {
+  filter {
+    name   = "tag:Name"
+    values = ["${var.supporting_resources_name}*.pri.*"]
+  }
+}
+
+data "aws_subnet" "private" {
+  for_each = toset(data.aws_subnets.private.ids)
+  id       = each.value
+}
+
 data "aws_elb_service_account" "main" {}
 
 ### LB Bucket Policy
@@ -62,5 +74,15 @@ data "aws_iam_policy_document" "s3" {
       identifiers = ["delivery.logs.amazonaws.com"]
       type        = "Service"
     }
+  }
+}
+
+data "aws_ami" "amazon_linux" {
+  most_recent = true
+  owners      = ["amazon"]
+
+  filter {
+    name   = "name"
+    values = ["amzn2-ami-kernel-5.10-*-${var.architecture}-gp2"]
   }
 }